Changed dynamic graph to start as a static graph
Changed initial representation of DynamicGraphView from a static Canvas
to a StaticGraphView representation of the data.
Added padding to the DynamicGraphView for usability when dismissing the
sheet.
Added a border to DynamicGraphView to hint at interactivity.
Fixed ScrollView scrolling usability by using TapGesture isntead of
LongPressGesture.
Added StaticGraphView based on refactor of DynamicGraphView with less
initialization and no gestures.