Added human friendly ranges
Added a readable range to calculate ranges that are friendly to humans
reading them.
Changed dynamic and static graph views to make use of a readable range
for the y-axis labels.
Changed the axis origin placeholder to use a dynamic number of digits.
Refactored the y-axis into its own view.