Changed readable pace range to handle zero speeds
Changed ReadablePaceRange init function to be optional so that it can
return nil for zero speeds that result in infinite pace and a runtime
crash.
Added an assertAlmostEqual function for unit tests where floating point
values are compared and may not match exactly.
Changed readable pace range unit tests to comply with optional
initializer.
Added a unit test for readable pace ranges with a zero speed parameter.